requirements.txt
ag-ui-protocol==0.1.8
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anthropic==0.60.0
anyio==4.10.0
argcomplete==3.6.2
attrs==25.3.0
boto3==1.40.2
botocore==1.40.2
cachetools==5.5.2
certifi==2025.8.3
charset-normalizer==3.4.2
click==8.2.1
cohere==5.16.2
colorama==0.4.6
distro==1.9.0
eval_type_backport==0.2.2
executing==2.2.0
fastapi==0.116.1
fastavro==1.12.0
filelock==3.18.0
frozenlist==1.7.0
fsspec==2025.7.0
google-auth==2.40.3
google-genai==1.28.0
googleapis-common-protos==1.70.0
griffe==1.9.0
groq==0.30.0
h11==0.16.0
hf-xet==1.1.5
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.34.3
idna==3.10
importlib_metadata==8.7.0
Jinja2==3.1.6
jiter==0.10.0
jmespath==1.0.1
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
logfire==4.1.0
logfire-api==4.1.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mcp==1.12.3
mdurl==0.1.2
mermaid-py==0.8.0
mistralai==1.9.3
multidict==6.6.3
openai==1.98.0
opentelemetry-api==1.36.0
opentelemetry-exporter-otlp-proto-common==1.36.0
opentelemetry-exporter-otlp-proto-http==1.36.0
opentelemetry-instrumentation==0.57b0
opentelemetry-proto==1.36.0
opentelemetry-sdk==1.36.0
opentelemetry-semantic-conventions==0.57b0
packaging==25.0
prompt_toolkit==3.0.51
propcache==0.3.2
protobuf==6.31.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.7
pydantic-ai==0.5.0
pydantic-ai-slim==0.5.0
pydantic-evals==0.5.0
pydantic-graph==0.5.0
pydantic-settings==2.10.1
pydantic_core==2.33.2
Pygments==2.19.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.4
rich==14.1.0
rpds-py==0.26.0
rsa==4.9.1
s3transfer==0.13.1
six==1.17.0
sniffio==1.3.1
sse-starlette==3.0.2
starlette==0.47.2
tenacity==8.5.0
tokenizers==0.21.4
tqdm==4.67.1
types-requests==2.32.4.20250611
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
uvicorn==0.35.0
wcwidth==0.2.13
websockets==15.0.1
wrapt==1.17.2
yarl==1.20.1
zipp==3.23.0
